pyproject.toml
[tool.poetry]
name = "pyprotogen"
version = "0.0.1"
description = ""
authors = ["Sergio Soldatov <soldatovsr97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.4.1"
pyyaml = "^6.0"
jinja2 = "^3.1.1"
rich = "^12.2.0"
inflection = "^0.5.1"
grpcio-tools = "1.51.1"
py-grpc-prometheus = {git = "https://github.com/sollof/py-grpc-prometheus.git", rev = "4932905"}
orjson = "^3.8.7"
mypy-protobuf = "^3.4.0"

[tool.poetry.scripts]
pyprotogen = 'pyprotogen.main:run'

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
importlib-metadata = "^4.11.4"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "env",
]
line-length = 100
